new function and error corrections
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/kebabs@110019 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/GappyPairC.cpp
@@ -4364,7 +4364,7 @@ void getWeightedFeatOfSVGappyPair(T maxUnSignedIndex, SEXP **pdFeatWeights, khas
                 }
 
                 kmer[index--] =
-                    alphaInf->reverseIndexMap[(int)((featIndex % (int) powAlpha[j + 1]) / powAlpha[j])];
+                    alphaInf->reverseIndexMap[(int)((featIndex % powAlpha[j + 1]) / powAlpha[j])];
             }
 
             SET_STRING_ELT(rownames, row, Rf_mkChar(kmer));
